3/requirements.txt
redis
flask
python-dotenv
